Backend/requirements.txt
langchain==0.0.284
python-dotenv==1.0.0
tiktoken==0.4.0
faiss-cpu==1.7.4
google-generativeai
protobuf~=3.19.0
psx-data-reader==0.0.6
Flask
Flask-Cors
requests
pandas
gunicorn